Ignore warning 44 (open shadowing)
Warning 44: "Open statement shadows an already defined identifier."
This warning happens on every example because we need to have the
user `open BsPuppeteer` if they actually want to use this code,
but because the module is already available we get a warning.
Should be fine to ignore but we should be careful about opening
any other libraries later.