Rename Target.pageType to targetType
In fact only one of the targets is a page, so this is more
appropriately called targetType. I prefer this over using type_
with the underscore to avoid the reserved word, and this way it
does not use the same name as the type_ function.

Adds deprecations on pageType and its bs.deriving functions.