Refactor main evaluation functions into Evaluator
The evaluation functions exploded in number because of all the type
cases and number of arguments that we needed to cover. The least
we can do is only define them once. This changeset creates the
Evaluator module, which defines an implementation for the set of
evaluate functions. This is the minimum set of evaluation functions
that any Puppeteer object has. Previously these externals were
repeated in ExecutionContext and FrameBase. They will be included
in the forthcoming Workers API as well.