Skip tests requiring the network
Downloading a whole copy of chromium every time we run the tests is too
inconsistent and slow. We can test some parts of BrowserFetcher with the
revision that is downloaded during the Puppeteer install, and those
tests are kept. We could maybe use some flag or env variable to
optionally run them, but it doesn't seem worth the effort at the moment.

This commit sponsored by a faulty modem and being forced to tether
to my phone.