fix: evaluateHandle return types
Return types were wrong for the eval functions in the
Evaluator.evaluateHandle family of functions. The eval function
should return some JS object and it's Puppeteer's job to convert
that to a handle. We had the user-provided eval function returning
a handle which they don't have a way of constructing. This was
missed because of using `%raw` for the eval function in the tests
with type annotations, so a wrong type was inferred.

While looking over this function set I also noticed that some had
the whole function's return type as `'r` rather than `JSHandle.t`.
We know for sure these all return a `JSHandle.t` or promise of one.
These types are also corrected.