Retool search example to use a different site
This broke at some point recently due to changes in the html and
upstream changed the example due to it being an example of a Google TOS
violation. Change it to search the Google developers site and do
something similar with the results. This does a slightly simplified
version of what upstream now does for their search example now, but in
Reason.